fix: refacotor project structure
