--- src/state/state.ts
@@ -1,12 +1,0 @@
-// import { cart } from "../components/templates/types";
-
-// class State {
-//     view: string
-//     cart: cart
-//     constructor() {
-//         this.view ='grid'
-//         this.cart = { items: []}
-//     }
-// }
-
-// export default State;
